refactor interfaces
