+++ src/client/src/store/azureProfileData/appService/reducer.ts
@@ -4,6 +4,7 @@ import AzureActionType from "../azureActionType";
 import { CONFIG_TYPEKEYS } from "../../typeKeys";
 import WizardInfoType from "../../templates/templatesType";
 import { IAppService } from "./model";
+import RootAction from "../../ActionType";
 
 const initialState: IAppService = {
   siteNameAvailability: {
@@ -18,7 +19,7 @@ const initialState: IAppService = {
 
 const appServiceReducer = (
   state: IAppService = initialState,
-  action: AzureActionType | WizardInfoType
+  action: AzureActionType | WizardInfoType | RootAction
 ) => {
   switch (action.type) {
     case AZURE_TYPEKEYS.SET_APP_NAME_AVAILABILITY:
