+++ src/client/src/store/templates/pages/action.ts
@@ -3,13 +3,13 @@ import { TEMPLATES_TYPEKEYS } from "../../typeKeys";
 import { IOption } from "../../../types/option";
 import { IApiTemplateInfo } from "../../../types/apiTemplateInfo";
 
-const getPagesOptionsAction = (pagesOptions: IOption[]) => {
-  return getPagesOptionsSuccessAction(
+const setPagesOptionsAction = (pagesOptions: IOption[]) => {
+  return setPagesOptionsSuccessAction(
     getOptionalFromApiTemplateInfo(getApiTemplateInfoFromJson(pagesOptions))
   );
 };
 
-const getPagesOptionsSuccessAction = (
+const setPagesOptionsSuccessAction = (
   pagesOptions: IOption[]
 ): IPageOptionsActionType => ({
   payload: pagesOptions,
@@ -50,6 +50,6 @@ function getOptionalFromApiTemplateInfo(items: IApiTemplateInfo[]): IOption[] {
 }
 
 export {
-  getPagesOptionsAction,
-  getPagesOptionsSuccessAction
+  setPagesOptionsAction,
+  setPagesOptionsSuccessAction
  };
