--- src/client/src/store/templates/preview/previewReducer.ts
@@ -1,16 +1,0 @@
-import WizardContentActionType from "../templatesActionType";
-import { CONFIG_TYPEKEYS } from "../../typeKeys";
-
-const previewStatus = (
-  state = false,
-  action: WizardContentActionType
-) => {
-  switch (action.type) {
-    case CONFIG_TYPEKEYS.SET_PREVIEW_STATUS:
-      return action.payload;
-    default:
-      return state;
-  }
-};
-
-export default previewStatus;
