+++ src/client/src/store/templates/templatesActionType.ts
@@ -1,12 +1,10 @@
 import { IUpdateFrameworkActionTypeAction, IFrontendFrameworksActionTypeAction, IBackendFrameworksSuccessActionTypeAction } from "./frameworks/model";
 import { IPageOptionsActionType } from "./pages/model";
-import { IPreviewStatusActionTypeAction } from "./preview/model";
 
 type WizardContentActionType =
   | IBackendFrameworksSuccessActionTypeAction
   | IFrontendFrameworksActionTypeAction
   | IUpdateFrameworkActionTypeAction
-  | IPageOptionsActionType
-  | IPreviewStatusActionTypeAction;
+  | IPageOptionsActionType;
 
 export default WizardContentActionType;
