+++ src/client/src/store/typeKeys.ts
@@ -2,7 +2,6 @@ export enum WIZARD_INFO_TYPEKEYS {
   SET_VISITED_WIZARD_PAGE = "WTS/wizardNavigation/SET_VISITED_WIZARD_PAGE",
   RESET_VISITED_WIZARD_PAGE = "WTS/wizardNavigation/RESET_VISITED_WIZARD_PAGE",
   SET_PAGE_WIZARD_PAGE = "WTS/wizardNavigation/SET_PAGE_WIZARD_PAGE",
-  SET_DETAILS_PAGE_INFO = "WTS/details/SET_DETAILS_PAGE_INFO",
   UPDATE_TEMPLATE_GENERATION_STATUS_MESSAGE = "WTS/postgen/UPDATE_TEMPLATE_GENERATION_STATUS_MESSAGE",
   UPDATE_TEMPLATE_GENERATION_STATUS = "WTS/postgen/UPDATE_TEMPLATE_GENERATION_STATUS",
   UPDATE_DEPENDENCY_INFO = "WTS/dependency/UPDATE_DEPENDENCY_INFO",
